+++ Project2.app/DataAccess/SpellRepo.cs
@@ -5,7 +5,6 @@ namespace Project2.app.DataAccess;
 
 public class SpellRepo(ApplicationDbContext context) : IRepo<Spell>
 {
-
     private readonly ApplicationDbContext _context = context;
 
     public async Task<Spell> CreateEntity(Spell spell)
@@ -15,9 +14,10 @@ public class SpellRepo(ApplicationDbContext context) : IRepo<Spell>
         return spell;
     }
 
-    public async Task<Spell> DeleteEntity(int id)
+    public async Task<Spell?> DeleteEntity(int id)
     {
-        Spell toDelete = _context.Spells.Find(id);
+        var toDelete = await GetById(id);
+        if (toDelete is null) return null;
         _context.Spells.Remove(toDelete);
         await _context.SaveChangesAsync();
         return toDelete;
@@ -28,22 +28,14 @@ public class SpellRepo(ApplicationDbContext context) : IRepo<Spell>
         return await _context.Spells.ToListAsync();
     }
 
-    public async Task<Spell> GetById(int id)
+    public async Task<Spell?> GetById(int id)
     {
         return await _context.Spells.FirstOrDefaultAsync(t => t.SpellId == id);
     }
 
-    /* public Task<Player> UpdateEntity(int id, Player updatePlayer)
+    public async Task<Spell?> UpdateEntity(int id, Dictionary<string, object> updates)
     {
-        Player oldPlayer = _context.Players.Find(id)!;
-        oldPlayer.Health = updateTrainer.Username;
-        oldPlayer.Level = updateTrainer.Password;
-        await _context.SaveChangesAsync();
-        return oldTrainer;
-    } */
-    public async Task<Spell> UpdateEntity(int id, Dictionary<string, object> updates)
-    {
-        Spell originalSpell = _context.Spells.FirstOrDefault(p => p.SpellId == id);
+        var originalSpell = await GetById(id);
 
         if (originalSpell != null)
         {
@@ -56,8 +48,8 @@ public class SpellRepo(ApplicationDbContext context) : IRepo<Spell>
                 }
 
             }
+            await _context.SaveChangesAsync();
         }
-        await _context.SaveChangesAsync();
-        return originalSpell;
+        return null;
     }
 }
